Refactor WebcastType to own class
