+++ android/src/main/java/com/thebluealliance/androidclient/helpers/WebcastHelper.java
@@ -7,65 +7,30 @@ import android.net.Uri;
 import com.google.gson.JsonObject;
 import com.thebluealliance.androidclient.R;
 
-/**
- * Created by phil on 3/27/15.
- */
 public class WebcastHelper {
 
-    public enum TYPE {
-        YOUTUBE,
-        TWITCH,
-        USTREAM,
-        LIVESTREAM,
-        IFRAME,
-        HTML5,
-        STEMTV,
-        NONE;
-
-        public String render(Context context) {
-            switch (this) {
-                case YOUTUBE:
-                    return context.getString(R.string.webcast_type_youtube);
-                case TWITCH:
-                    return context.getString(R.string.webcast_type_twitch);
-                case USTREAM:
-                    return context.getString(R.string.webcast_type_ustream);
-                case LIVESTREAM:
-                    return context.getString(R.string.webcast_type_livestream);
-                case IFRAME:
-                    return context.getString(R.string.webcast_type_gameday); // watch on web GameDay
-                case HTML5:
-                    return context.getString(R.string.webcast_type_html5);
-                case STEMTV:
-                    return context.getString(R.string.webcast_type_stemtv);
-                default:
-                    return "";
-            }
-        }
-    }
-
-    public static TYPE getType(String typeString) {
+    public static WebcastType getType(String typeString) {
         switch (typeString) {
             case "youtube":
-                return TYPE.YOUTUBE;
+                return WebcastType.YOUTUBE;
             case "twitch":
-                return TYPE.TWITCH;
+                return WebcastType.TWITCH;
             case "ustream":
-                return TYPE.USTREAM;
+                return WebcastType.USTREAM;
             case "livestream":
-                return TYPE.LIVESTREAM;
+                return WebcastType.LIVESTREAM;
             case "iframe":
-                return TYPE.IFRAME;
+                return WebcastType.IFRAME;
             case "html5":
-                return TYPE.HTML5;
+                return WebcastType.HTML5;
             case "stemtv":
-                return TYPE.STEMTV;
+                return WebcastType.STEMTV;
             default:
-                return TYPE.NONE;
+                return WebcastType.NONE;
         }
     }
 
-    public static Intent getIntentForWebcast(Context context, String eventKey, TYPE type, JsonObject params, int number) {
+    public static Intent getIntentForWebcast(Context context, String eventKey, WebcastType type, JsonObject params, int number) {
         switch (type) {
             case YOUTUBE:
                 return getWebIntentForUrl(context.getString(R.string.webcast_youtube_embed_pattern, params.get("channel").getAsString()));
